generalize getColumns
Make it so that we can call getColumns with different
formatters and transforms. This gives us a standard
template for the columns object with a small amount
of flexibility.

I'm planning on using this for the "Add new row" button.
These could get moved into separate components at some
point, but I don't think it's necessary yet.