Setup connectDragSource on row children(cells), Use connectDragPreview on row
- This makes it so connectDragPreview and the row's children are passed
into DraggableRow.

Within the React.createElement call we don't call connectDragSource for the
entire row but for each individual child cell, making it so the row can
be dragged from any of a row's cell (this doesn't necessarily change
behavior but does allow us to have finer control over which parts of a
row can start a drag). We also add a call to connectDragPreview for the entire
row so that despite wherever the drag is started from for a row, the preview
shown to the user will still be of the entire thing i.e. dragging by a handle
won't show a preview of just the handle being dragged.