Use draggable-row.jsx, Only connectDragSource on row handle
- This makes it so our table uses the draggableRow we define in
draggable-row.jsx.

A DndRow component is created that is essentially the same as
the previous dnd.Row except for connectDragSource being set for each of
the child cells rather than the entire row.

A DndCell component is created that contains a render function that
removes the connectDragSource from the props of the component and
returns a td cell. 'connectDragSource' is only called on cells that
contain the drag-handle-cell class that identify it as the drag handle.